Add TFS support for VS2017 users.
Visual Studio 2017 changes the on-disk format of the user cache for TFS
workspaces, and Team Explorer Everywhere and the TFS Java SDK don't have
support for the new format. This meant that our two methods for interfacing
with TFS on the client would not work against VS2017 workspaces.

Fortunately, the `tf.exe` command-line tool that ships with VS2017 restores
some of the features that were removed in VS2015 (which prompted our move to
Team Explorer Everywhere), and is sufficient for posting pending changes. This
change implements support for that, and adds some documentation about the three
TFS methods and how to use them.

Testing done:
- Posted a few TFS changes involving edits, adds, deletes, moves, and copies on
  various types of files, including binary files.
- Built the docs and looked at the new page.

Reviewed at https://reviews.reviewboard.org/r/8833/