Get rid of die().
The `die()` method is one of the most ancient parts of RBTools, and it's long
outlived its usefulness, especially now that more and more people are using
RBTools as an API and not just a command.

This change gets rid of it in favor of using exceptions, which are already all
handled correctly.

Testing done:
- Smoke tested rbt commands.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8899/