Use `open <url>` on macOS instead of webbrowser module.
There's currently a bug in the interaction between AppleScript on the latest
macOS and python's `webbrowser` module which is causing URL opens to fail. This
change moves us over to using macOS's `open` command instead, which is likely
the way that Python is going to go.

Testing done:
Posted this change and opened it successfully.

Reviewed at https://reviews.reviewboard.org/r/8959/