Improve the RBTools doc builds and fix some warnings and issues.
The RBTools documentation and release notes builds now use the
intersphinx linking setup we use in our other doc codebases. There are
intersphinx entries for different versions of Review Board and RBTools,
which the release notes and manual will later be able to take advantage
of, and the doc builds can be set to point to a local website instance
to find the object inventories.

There's also fixes for a bad directive name and for missing escaping for
RBTools command help output.

Testing Done:
Built the docs against the local webserver and against reviewboard.org,
and saw that the intersphinx URLs matched what I pointed to.

Saw that errors relating to the escaping and bad directive were gone.

Reviewed at https://reviews.reviewboard.org/r/8950/