Improve query optimization when using ClearCase-based repositories.
We optimize HTTP requests to Review Board by prioritizing repositories
based on name. Previously this was done using strict comparison of name
returned by Review Board to name returned by cleartool. This is no good
when supporting users on Windows and Unix both. They are different, as
shown here:

    W:\srcSHARE>cleartool describe -short vob:.
    \srcSHARE

And when the VOB is mounted to /vobs in Unix:

    [/vobs/srcSHARE]$ cleartool describe -short vob:.
    /vobs/srcSHARE

Now we compare the name of the returned repository with each part of
the VOBs tag returned by cleartool. Subsequently, Review Board server
administrators can set repository names for individual VOBs and can
expect it to match for users regardless of platform. `srcSHARE` works
in the example above.

Testing Done:
I posted a review to an internally-hosted Review Board instance from a
VOB with the name `srcSHARE` on both Linux and Windows. The repository
in Review Board has the name `srcSHARE`.

In the debug output it is clear that the number of repositories queried
is reduced. In our specific case because the Review Board server is
hosted very far away, this represents a time savings of a few minutes
for each review posted.

More details are in the review request.

Reviewed at https://reviews.reviewboard.org/r/8936/