Clarify the default tracking branch behavior in Git.
The help for --tracking-branch said that if not provided, origin/master
would be used as a default. This is sort of true, but misses a step.
RBTools will attempt to find the tracking branch for the active branch
and use that instead. This wasn't really clear from the help text, and
led to some confusion by users.

This is a simple wording change that just adds information on the
missing step.

Testing Done:
Saw the help text.

Reviewed at https://reviews.reviewboard.org/r/8965/