Prevent VS2015 tf.exe from taking priority over rb-tfs.
VS2015 ships a tf.exe, but it isn't full-featured enough for us to use.
Unfortunately, when I added support for VS2017's version, I did so in a way
that prevented rb-tfs from being used. This change fixes it up to actually
check the version of tf.exe before assuming that it will work.

Testing done:
Had a customer use this in production and they reported that it worked.

Reviewed at https://reviews.reviewboard.org/r/9012/