Don't include "svn info" output in logging.
A user was running into a problem where rbt was crashing with a
UnicodeError inside the SVN client when attempting to detect the
repository type. In their case, they weren't even using subversion, so
it was particularly silly.

While this can be worked around by setting REPOSITORY_TYPE, there's no
reason to log the "svn info" output during repo detection--either the
command returns success or error, and if it's an error, we know that
it's not a (working) subversion checkout. Not logging the output
insulates us from possible locale issues.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/9370/