Make sure to load zip file as binary when installing things.
On most systems this doesn't matter, but we did see one case in the wild
where the hash of the downloaded zip file wasn't matching, which I
believe was caused by some newline munging. This change makes sure to
load the downloaded zip file as binary to compute the hash.

Testing done:
Ran "rbt install" successfully. Of course, it never failed for me to
start with.

Reviewed at https://reviews.reviewboard.org/r/9372/