Log size information for generated diffs when posting a change.
This adds some additional logging after generating a diff. The log
entries show the size of the diff and the parent diff (if generated) in
bytes. This should help when gathering debug information from users who
are hitting errors uploading diffs.

Testing Done:
Posted a diff to a server and saw the size information for both the
diff and parent diff.

Reviewed at https://reviews.reviewboard.org/r/9428/