Fix development dependencies for Python 2.6.
For development, RBTools depends on `hgsubversion`, which depends on any
version of `mercurial`. However, the latest version of `mercurial` does
not work on Python 2.6.

This change explicitly lists `mercurial` as a dependency and adds
version specifiers to cap the version on Python 2.6.

Testing Done:
Tested installation on Python 2.6 and 2.7.

Reviewed at https://reviews.reviewboard.org/r/9527/