Fix up a variety of unicode/bytes issues with SVN.
The SVN client was handling a variety of things using the wrong string
types, most notably filenames. In Python 3, filenames are expected to be
treated as `str` internally, and then serialized/deserialized using the
filesystem encoding (theoretically this is also the case in Python 2,
but the implicit casting makes it easy to be sloppy). This change fixes
those up so that we're at least sucessfully running the tests.

This also incidentally fixes bug 4547, which involves using
`Popen.communicate()` instead of `Popen.wait()`.

Testing done:
Ran unit tests on Python 2.7 and 3.6.

Fixes bug 4547.

Reviewed at https://reviews.reviewboard.org/r/9627/