Fix up some more unicode vs. bytes issues in RBTools.
There were a few more places where we had the wrong types in use, this
time related to the password prompts and the creation of HTTP headers
(especially related to Basic auth credentials). This change fixes those
up.

Testing done:
Was able to log in and post changes to both the devserver and public
servers using Python 2.7 and 3.6.

Reviewed at https://reviews.reviewboard.org/r/9641/