Fix normalization of filenames for -X with Git.
The Git client normalizes the exclude patterns to all be absolute paths,
but it was then passing in the wrong base directory for joining with the
actual filenames being tested. This change fixes it to use the same base
directory for both.

Testing done:
Posted a change with -X and saw the files actually get excluded.

Reviewed at https://reviews.reviewboard.org/r/9715/