Update development dependencies and move doc-specific ones out.
This updates the required versions of kgb, Sphinx, and beanbag-docutils
to the latest versions.

Along with this, it moves the doc-related dependencies into
doc-requirements.txt, so they're not needed for building the source
tree. This helps with build issues on Python 2.6 and just lowers the
requirements overall for working on the source tree.

Testing Done:
Successfully installed both the development and doc dependencies.

Unit tests pass (except for two that are broken upstream).

Reviewed at https://reviews.reviewboard.org/r/9833/