Fix GitClient.get_repository_info to not change the working directory.
One particularly annoying aspect of the GitClient is that it changes the
working directory after scanning. This has a variety of impacts, and
makes it hard to implement certain things like more advanced client
detection.

This change fixes it up so that we can pass in a cwd argument to
`Popen`. Instead of changing the working directory for the entire
process, we just do so for the subprocesses that get executed.

Testing done:
- Ran unit tests on Python 2.7 and 3.6.
- Did a variety of operations with a git repository.

Reviewed at https://reviews.reviewboard.org/r/9886/