Avoid TypeError while coercing string to int
A change in the `rbtools patch` support a while back regressed the
behaviour of the `-p` flag that is passed to patch(1). We now no longer
assume that is an `int` and just splat it into the string to avoid a
TypeError when `p` is a string like `'4'`.

Testing Done:
Ran unit tests.

Bugs closed: 4364

Reviewed at https://reviews.reviewboard.org/r/9940/