Add status-update command to rbtools.
Review Board 3.0 introduced new support for "status updates" which are used to
indicate if a review request is undergoing some automated checks of some kind
(for instance, automated code review or pending builds). These can have a state
(such as "pending", "error", "success", and others), a summary, optional link,
and a review (when there are errors).

This adds a new command for working with status updates, helping to
create/update the state of a status update. This command can be run with `rbt
status-update`.

Testing Done:
Debug testing of the `--review` input on local installation with the following
file contents:

```
{
    "review": {
        "body_top": "Header comment"
    },
    "diff_comment": [
        {
            "filediff_id": 10,
            "first_line": 729,
            "issue_opened": true,
            "num_lines": 1,
            "text": "Adding a comment on a diff line",
            "text_type": "markdown"
        }
    ],
    "general_comments": [
        {
            "text": "Adding a general comment",
            "text_type": "markdown"
        }
    ]
}
```

Reviewed at https://reviews.reviewboard.org/r/9343/