Fix and improve doc building for the RBTools manual.
Our custom Sphinx extension for RBTools was not compatible with modern
versions of Sphinx, due to a bad import. This is a pretty trivial fix by
itself, and one we've already done for Review Board, but this change
also goes a step further and adds some of the beanbag-docutils
extensions needed to enhance references and to link to integrations on
the website.

Testing Done:
Built the docs. Checked that the integration link worked.

Reviewed at https://reviews.reviewboard.org/r/10034/