Clean up git client type.
When we detect what type of git checkout we're dealing with (plain git,
git-svn, or git-p4), that would get assigned to a public attribute with a
string type. This change fixes things so it's a private attribute with
symbols instead.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/10086/