Clean up computation of upstream branch and origin.
The git client in RBTools had a bunch of crufty code to determine the
name of the remote branch and the origin URL/remote name. This change
cleans those up to be simpler and more consistent.

The `self._upstream_branch` attribute still isn't used for the tracking
branch--that will come in the next change.

Testing done:
- Posted some changes.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/10089/