Fix rbt patch --print on Python 3.
On Python 3, the data returned from the API is a byte array, not a Unicode
string. As such, it has to be decoded before it can be passed to the print()
function. Otherwise, the print() function will print the representation of the
byte array. See the bug for details.

Testing Done:
- Ran `rbt patch --print` under Python 2.7 and 3.6.
- Ran `rbt patch` _without_ `--print` under Python 2.7 and 3.6.
- Ran unit tests under Python 2.7 and 3.6.

Bugs closed: 4736

Reviewed at https://reviews.reviewboard.org/r/10140/