Post review requests with history when supported
When the Review Board server has the `review_request.supports_history`
capability, we will default to posting with history. Using history can
be forced with the new `-H/--with-history` flag, which will cause an
error if the server does not support it.

Traditional-style squashed review requests are still supported: they can
be posted with `-S/--squash-history`.

Review requests that were created with history will *always* be updated
with history and review requests not created with history will,
likewise, *never* be updated with history. Since review requests in
Review Board exist in one state or the other, attempting to post would
result in an error, so we do not let users attempt such posts.

If the server does not support history, the `-S` and `-H` options are
ignored.

Testing Done:
Posted this review request.
Posted a review request with multiple commits to my devserver.

Reviewed at https://reviews.reviewboard.org/r/9911/