Raise error when no history entries present or a history entry is empty
Previously, the case where `SCMTool.get_commit_history` returned `None`
was not graciously handled in `rbt post` leading to an exception. We now
explicitly handle this case and report this to the user.

Additionally, we now explicitly handle the case of a commit containing
an empty diff (e.g., if the commit was created with
`git commit --allow-empty`), which will not pass validation with Review
Board (as it doesn't allow empty diff uploads).

Testing Done:
- Attempted to post an empty history and the appropriate error was
  shown.
- Attempted to post a history with an empty commit and the
  appropriate error was shown.

Reviewed at https://reviews.reviewboard.org/r/10281/