Fix inclusion of binary file changes in Perforce diffs.
The diff code for Perforce clients was mistakenly filtering out changes
to binary files. This change fixes it.

Testing done:
Created a perforce client with a change to a binary file. Before,
posting the changeset would show only the text files. After, it shows
the expected "This is a binary file" line in the diff viewer.

Fixes bug 3659.

Reviewed at https://reviews.reviewboard.org/r/10326/