Fix posting review requests using -u.
With the DVCS work, some of the logic for updating review requests
changed ever so slightly, breaking the updating of review requests that
didn't already have a draft.

Prior to this work, the process of guessing review requests involved
making queries with minimal information, just enough to match up the
contents of review requests, and once found, we'd throw away the matched
review request and re-fetch using its ID.

The DVCS work subtly changed this by simply reusing the review request
object, which meant that some information wasn't present. This included
the link to the draft resource, which was expanded (and possibly empty)
during the guessing process. While it saved an API request, it meant
that only review requests with drafts could ever be updated.

This change simply reverts that bit of logic back to how it used to be,
with a comment detailing why we need to re-fetch the review request.

Testing Done:
Posted a review request against reviews.reviewboard.org and then updated
it successfully.

Reviewed at https://reviews.reviewboard.org/r/10381/