Improve the error details when accessing missing resource attributes.
When attempting to access an attribute that doesn't exist on a resource,
the caller would just get a plain `AttributeError`. Left uncaught, this
would be displayed to the user, making it very hard to report anything
meaningful.

This change adds explicit error information containing the attribute
being requested and the resource it's being requested on.

Testing Done:
Used this to diagnose a separate problem. Saw the enhanced error details.

Reviewed at https://reviews.reviewboard.org/r/10383/