Fix regressions posting changes for review with custom diff files.
Some of the DVCS work broke posting changes using `--diff-filename`
for review. There was a typo in an argument to a method, and a
variable that didn't have a default value. This is a simple change
that fixes both.

Testing Done:
Posted several changes on my dev server with this.

Reviewed at https://reviews.reviewboard.org/r/10460/