Add release notes for RBTools 1.0.2.
This covers a large number of Python 3.x fixes, as well as a reworking
of the Two-Factor Authentication support for RBCommons, helpful
instructions when RBTools is pointing at a bad URL, better smarts for
Git repositories not using an `origin` remote, a flag to turn off
renames for Git, and more.

Testing Done:
Built the docs. Checked for spelling, build, and link errors.

Reviewed at https://reviews.reviewboard.org/r/10540/