Add documentation on available RBTools environment variables.
RBTools 1.0.3 is introducing the `RBTOOLS_EDITOR` environment variable,
for customizing the editing experience. This is one of a few environment
variables we offer, none of which are documented anywhere.

This change introduces a section to the User Configuration page to list
available environment variables. Any `:envvar:` references to these will
link to the correct place.

Testing Done:
Built the docs. Checked for build and spelling errors.

Checked that each of the listed environment variables rendered correctly
and could be linked to.

Reviewed at https://reviews.reviewboard.org/r/10965/