Fix Subversion path normalization on Python 3.7+.
The code to split a path and normalize repeated slashes for Subversion
URLs was using a regex that wasn't very safe. It was splitting on `/*`,
which means 0 or more slashes. Python 3 warned about this, but kept
behavior the same. Python 3.7 removed the backwards-compatibility, and
the result was that every single character was split into its own item
in an array.

This fixes the regex to be `/+`, which is the correct form of what we
wanted, fixing compatibility with all Python versions.

Testing Done:
Unit tests pass on all supported versions of Python.

Reviewed at https://reviews.reviewboard.org/r/10995/