Make "rbt patch" use the review request details when there's one commit.
When using "rbt patch" to apply a change from a review request that was
created with history, it uses the commit message as the message for the
new commit that it's creating. This is ideal when applying a review
request that has multiple commits (since it pulls in all of the messages
for each commit in turn), but it didn't feel right for the simpler case
of applying a change with a single commit. In order to maintain the same
behavior as before, we now use the review request details as the commit
message when there's only one commit.

Testing Done:
Verified that after this change, `rbt patch -c` used the review request
details as the commit message instead of the commit metadata.

Reviewed at https://reviews.reviewboard.org/r/11019/