Allow the Mercurial client to use a custom 'hg' executable.
Mercurial provides a command-server for faster operations. This is used
through the `chg` binary, rather than `hg`.

This change allows custom scripts to pass in `chg` (or any other
executable) when constructing `MercurialClient`. It does not, however,
enable it for standard RBTools usage at this time.

This has sped up a single post from 2.8 seconds to 1.1 seconds in
the deployment in which this was developed and tested.

Testing Done:
Used `MercurialClient('chg')` in a custom script and saw that debugging
output calls `chg` instead of `hg`.

Verified thed default `MercurialClient()` still used the `hg` binary.

Reviewed at https://reviews.reviewboard.org/r/11102/